requirements.txt
pandas>=1.0.0
jieba>=0.42.1
torch>=1.7.0
transformers>=4.0.0
numpy>=1.19.0
tqdm>=4.45.0
fastapi>=0.68.0
uvicorn>=0.15.0
gunicorn>=20.1.0
pyjwt>=2.1.0
